them tinh nang chinh sua product

+++ src/main/java/com/EcommerceShop/Shop/DTO/request/Product/UpdateProductDetailRequest.java
@@ -10,7 +10,7 @@ public class UpdateProductDetailRequest {
 
     private String info ;
 
-    private float price ;
+    private Double price ;
 
-    private int quantity ;
+    private Integer quantity ;
 }
